+++ lib/srec/input/file/stewie.cc
@@ -1,6 +1,6 @@
 //
 //      srecord - manipulate eprom load files
-//      Copyright (C) 1998-2008 Peter Miller
+//      Copyright (C) 1998-2009 Peter Miller
 //
 //      This program is free software; you can redistribute it and/or modify
 //      it under the terms of the GNU General Public License as published by
@@ -234,8 +234,8 @@ srec_input_file_stewie::read(srec_record &record)
                     fatal_error
                     (
                         "data record count mismatch (file %ld, read %ld)",
-                        addr,
-                        mask
+                        (long)addr,
+                        (long)mask
                     );
                 }
             }
