+++ lib/srec/input/filter/checksum.cc
@@ -1,6 +1,6 @@
 //
 //      srecord - manipulate eprom load files
-//      Copyright (C) 1998, 1999, 2001, 2002, 2006-2008 Peter Miller
+//      Copyright (C) 1998, 1999, 2001, 2002, 2006-2009 Peter Miller
 //
 //      This program is free software; you can redistribute it and/or modify
 //      it under the terms of the GNU General Public License as published by
@@ -77,7 +77,7 @@ srec_input_filter_checksum::read(srec_record &record)
     {
         if (width <= 1)
         {
-            for (int j = 0; j < record.get_length(); ++j)
+            for (size_t j = 0; j < record.get_length(); ++j)
             {
                 sum += record.get_data(j);
             }
@@ -85,7 +85,7 @@ srec_input_filter_checksum::read(srec_record &record)
         else if (end == endian_little)
         {
             // Little endian
-            for (int j = 0; j < record.get_length(); ++j)
+            for (size_t j = 0; j < record.get_length(); ++j)
             {
                 sum += (sum_t)record.get_data(j) << (8 *
                     ((record.get_address() + j) % width));
@@ -94,7 +94,7 @@ srec_input_filter_checksum::read(srec_record &record)
         else
         {
             // Big endian
-            for (int j = 0; j < record.get_length(); ++j)
+            for (size_t j = 0; j < record.get_length(); ++j)
             {
                 sum += (sum_t)record.get_data(j) << (8 *
                     (width - 1 - ((record.get_address() + j)
